rag/requirements.txt
tqdm
numpy
pandas 

fastapi 
uvicorn[standard]

requests
selenium
webdriver-manager 

transformers
huggingface_hub

faiss-cpu
scikit-learn
konlpy
kiwipiepy
rank-bm25

langchain
langchain_openai
langchain_upstage
langchain_community
langchain_huggingface